Adicionado arquivo conftest.py e criado o módulo main.py
